Update Makefile with "./" as the starting path, should produce same output.
The only change here is to add "./" to all the relative paths,
this is to make the next Makefile export from mbed-cli diff
much much smaller, concentrating only on the real changes.

The changes from this commit do not change the make output, it
should produce exactly the same builds.